pyproject.toml
[project]
name = "jirald-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.10.0",
    "boto3>=1.40.13",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "mcp>=1.13.0",
    "pygithub>=2.7.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
